+++ spec/fixtures/timeout.js
@@ -1,0 +1,5 @@
+'use strict'
+
+setTimeout(function () {
+  /* Just make the loop wait for some time */
+}, 3000)