+++ spec/spec-helpers.js
@@ -1,7 +1,7 @@
 'use babel'
 
 export function waitsForAsync(asynCallback, result) {
-  waitsForPromise(function () {
+  waitsForPromise({ timeout: 15 * 1000 }, function () {
     return asynCallback().then(function (returnValue) {
       if (typeof result !== 'undefined') {
         expect(returnValue).toEqual(result)
@@ -11,7 +11,7 @@ export function waitsForAsync(asynCallback, result) {
 }
 
 export function waitsForAsyncRejection(asynCallback, errorMessage) {
-  waitsForPromise(function () {
+  waitsForPromise({ timeout: 15 * 1000 }, function () {
     return asynCallback().then(function () {
       expect(false).toBe(true)
     }, function (error) {
@@ -21,3 +21,17 @@ export function waitsForAsyncRejection(asynCallback, errorMessage) {
     })
   })
 }
+
+// Jasmine 1.3.x has no sane way of resetting to native clocks, and since we're
+// gonna test promises and such, we're gonna need it
+function resetClock() {
+  for (const key in jasmine.Clock.real) {
+    if (jasmine.Clock.hasOwnProperty('real')) {
+      window[key] = jasmine.Clock.real[key]
+    }
+  }
+}
+
+beforeEach(function () {
+  resetClock()
+})